Moved get_accounts() & get_txs() into update.py
